+++ frontend/src/pages/Admin/CouponDesign/CustomCouponDesign/ChoiceTemplate/style.tsx
