+++ frontend/src/pages/Admin/CouponDesign/CustomCouponDesign/StampCustomModal/style.tsx
