+++ frontend/src/pages/Admin/CouponDesign/CustomCouponDesign/style.tsx
@@ -1,5 +1,5 @@
 import { styled } from 'styled-components';
-import { PageContainer } from '../../../style/layout/common';
+import { PageContainer } from '../../../../style/layout/common';
 
 export const CustomCouponDesignContainer = styled(PageContainer)`
   display: flex;
